+++ yt/visualization/tests/test_callbacks.py
@@ -117,10 +117,10 @@ def test_scale_callback():
         p.annotate_scale(corner="upper_right", coeff=10.0, unit="kpc")
         assert_fname(p.save(prefix)[0])
         p = SlicePlot(ds, "x", "density")
-        p.annotate_scale(text_args={"size": 24})
+        p.annotate_scale(text_kwargs={"size": 24})
         assert_fname(p.save(prefix)[0])
         p = SlicePlot(ds, "x", "density")
-        p.annotate_scale(text_args={"font": 24})
+        p.annotate_scale(text_kwargs={"font": 24})
         assert_raises(YTPlotCallbackError)
 
     with _cleanup_fname() as prefix:
@@ -364,7 +364,7 @@ def test_text_callback():
         # Now we'll check a few additional minor things
         p = SlicePlot(ds, "x", "density")
         p.annotate_text(
-            [0.5, 0.5], "dinosaurs!", coord_system="axis", text_args={"color": "red"}
+            [0.5, 0.5], "dinosaurs!", coord_system="axis", text_kwargs={"color": "red"}
         )
         p.save(prefix)
 
@@ -375,7 +375,7 @@ def test_text_callback():
         assert_raises(YTDataTypeUnsupported, p.save, prefix)
         p = ProjectionPlot(ds, "r", "density")
         p.annotate_text(
-            [0.5, 0.5], "dinosaurs!", coord_system="axis", text_args={"color": "red"}
+            [0.5, 0.5], "dinosaurs!", coord_system="axis", text_kwargs={"color": "red"}
         )
         assert_fname(p.save(prefix)[0])
 
@@ -580,7 +580,7 @@ def test_contour_callback():
             clim=(0.4, 0.6),
             mpl_kwargs={"linewidths": 2.0},
             label=True,
-            text_args={"fontsize": "x-large"},
+            text_kwargs={"fontsize": "x-large"},
         )
         p.save(prefix)
 
@@ -594,7 +594,7 @@ def test_contour_callback():
             clim=(0.4, 0.6),
             mpl_kwargs={"linewidths": 2.0},
             label=True,
-            text_args={"fontsize": "x-large"},
+            text_kwargs={"fontsize": "x-large"},
             data_source=s2,
         )
         p.save(prefix)
@@ -610,7 +610,7 @@ def test_contour_callback():
             clim=(1.0e-1, 1.0e1),
             label=True,
             mpl_kwargs={"colors": ("c", "w"), "linewidths": 1},
-            text_args={"fmt": "%1.1f"},
+            text_kwargs={"fmt": "%1.1f"},
         )
         assert_fname(slc.save(prefix)[0])
 
@@ -625,7 +625,7 @@ def test_contour_callback():
             clim=(0.4, 0.6),
             mpl_kwargs={"linewidths": 2.0},
             label=True,
-            text_args={"fontsize": "x-large"},
+            text_kwargs={"fontsize": "x-large"},
         )
         assert_raises(YTDataTypeUnsupported, p.save, prefix)